+++ examples/basic/src/entry-server.tsx
@@ -1,7 +1,1 @@
-export {
-  routes,
-  renderPage,
-  matchRoute,
-  resolveRouteModule,
-  RedirectResponse,
-} from "virtual:pages/server";
+export * from "virtual:pages/server";